Read time
I asked Twitter about this, and @yaroslavpat suggested getting rid of the word 'read' here. It makes sense, because any mention of reading here sounds too wordy and awkward, much worse than in English.

I also contracted 'minutes' to 'min.', because full word requires issue #55 to be resolved. It's okay in Russian, but it would be better to just pluralize minutes correctly.